Move onprem_utils
