+++ sky/backends/cloud_vm_ray_backend.py
@@ -32,8 +32,8 @@ from sky import task as task_lib
 from sky.data import data_utils
 from sky.data import storage as storage_lib
 from sky.backends import backend_utils
+from sky.backends import onprem_utils
 from sky.backends import wheel_utils
-from sky.onprem import onprem_utils
 from sky.skylet import autostop_lib
 from sky.skylet import job_lib
 from sky.skylet import log_lib
